refactor: 불필요객체(구매요청) 제거
수동로또개수, 자동로또개수 출력을 위한 정보를 특정 객체에 담지않고 바로 출력처리
