+++ README.md
@@ -165,16 +165,12 @@
   - [x] 당첨번호가 포함되어있는지 여부를 반환한다.
 - Lottos (일급콜렉션)
   - [x] 구매금액과 수동 로또 개수를 입력받아서 로또번호를 생성한다.
+  - [x] 수동로또를 구입할 금액은 총구매금액을 초과할 수 없다.
   - [x] 당첨번호와 비교하여 당첨결과를 구한다.
 - PurchaseAmount
   - [x] 로또를 사기위한 최소금액 여부 검증
   - [x] 구매금액으로 살 수 있는 로또 개수를 구한다.
   - [x] 수동로또개수만큼의 금액을 뺀 금액으로 살수 있는 로또 개수를 구한다.
-- PurchaseRequest
-  - [x] 구매요청정보(총 구매금액, 수동로또번호 콜렉션)를 가진다.
-  - [x] 수동로또를 구입할 금액은 총구매금액을 초과할 수 없다.
-  - [x] 구매할 수동로또개수를 구한다.
-  - [x] 구매할 자동로또개수를 구한다. (총 구매금액에서 수동로또 개수만큼 뺀 개수)
 - LottoNumber (원시값 포장)
   - [x] 로또 번호 속성을 가진다.
     - [x] 1~45 범위내의 숫자여야 한다.
