+++ src/test/java/lotto/domain/LottosTest.java
@@ -5,6 +5,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 @DisplayName("로또(일급콜렉션) 테스트")
 public class LottosTest {
@@ -18,8 +19,7 @@ public class LottosTest {
 	@Test
 	@DisplayName("로또를 모두 자동으로 구매하는 테스트")
 	void purchaseAutoLottos() {
-		PurchaseRequest purchaseRequest = new PurchaseRequest(new PurchaseAmount(14000));
-		Lottos purchaseLottos = new Lottos(randomNumbersGenerator, purchaseRequest.getPurchaseAmount(), purchaseRequest.getManualLottos());
+		Lottos purchaseLottos = new Lottos(randomNumbersGenerator, new PurchaseAmount(14000));
 		assertThat(purchaseLottos.size()).isEqualTo(14);
 	}
 
@@ -32,9 +32,20 @@ public class LottosTest {
 				new Lotto(1, 2, 3, 4, 5, 6),
 				new Lotto(1, 2, 3, 4, 5, 6)
 		);
-		PurchaseRequest purchaseRequest = new PurchaseRequest(new PurchaseAmount(14000), manaulLottos);
-		Lottos purchaseLottos = new Lottos(randomNumbersGenerator, purchaseRequest.getPurchaseAmount(), purchaseRequest.getManualLottos());
+		Lottos purchaseLottos = new Lottos(randomNumbersGenerator, new PurchaseAmount(14000), manaulLottos);
 		assertThat(purchaseLottos.size()).isEqualTo(14); //전체
+	}
+
+	@Test
+	@DisplayName("수동으로 구매할 로또개수가 총 구매금액을 초과하는 경우 예외발생 테스트")
+	void purchaseableAmountExceeded() {
+		Lottos manaulLottos = new Lottos(
+				new Lotto(1, 2, 3, 4, 5, 6),
+				new Lotto(1, 2, 3, 4, 5, 6),
+				new Lotto(1, 2, 3, 4, 5, 6)
+		);
+		assertThatThrownBy(() -> new Lottos(randomNumbersGenerator, new PurchaseAmount(2000), manaulLottos))
+				.isInstanceOf(IllegalArgumentException.class);
 	}
 
 }
